+++ spatial_access/src/dataFrame.h
@@ -12,38 +12,38 @@ public:
     std::vector<std::vector<unsigned short int>> dataset;
 
     // Initialization:
-    dataFrame(bool isSymmetric, unsigned int rows, unsigned int cols);
+    dataFrame(bool isSymmetric, unsigned long int rows, unsigned long int cols);
     
     // Getters:
-    unsigned short int getValueByLoc(unsigned int row_loc, unsigned int col_loc) const;
+    unsigned short int getValueByLoc(unsigned long int row_loc, unsigned long int col_loc) const;
     unsigned short int getValueById(const row_label_type& row_id, const col_label_type& col_id) const;
 
     bool getIsSymmetric() const;
-    unsigned int getRows() const;
-    unsigned int getCols() const;
+    unsigned long int getRows() const;
+    unsigned long int getCols() const;
     const std::vector<std::pair<col_label_type, unsigned short int>> getValuesByRowId(const row_label_type &row_id, bool sort) const;
     const std::vector<std::pair<row_label_type, unsigned short int>> getValuesByColId(const col_label_type &col_id, bool sort) const;
     const std::vector<row_label_type>& getRowIds() const;
     const std::vector<col_label_type>& getColIds() const;
-    const row_label_type& getRowIdForLoc(unsigned int row_loc) const;
-    const col_label_type& getColIdForLoc(unsigned int col_loc) const;
-    unsigned int getRowLocForId(const row_label_type& row_id) const;
-    unsigned int getColLocForId(const col_label_type& col_id) const;
-    const std::vector<unsigned short int>& getDatasetRow(unsigned int row) const;
+    const row_label_type& getRowIdForLoc(unsigned long int row_loc) const;
+    const col_label_type& getColIdForLoc(unsigned long int col_loc) const;
+    unsigned long int getRowLocForId(const row_label_type& row_id) const;
+    unsigned long int getColLocForId(const col_label_type& col_id) const;
+    const std::vector<unsigned short int>& getDatasetRow(unsigned long int row) const;
     const std::vector<std::vector<unsigned short int>>& getDataset() const;
 
     // Setters
-    void setValueByLoc(unsigned int row_loc, unsigned int col_loc, unsigned short int value);
+    void setValueByLoc(unsigned long int row_loc, unsigned long int col_loc, unsigned short int value);
     void setValueById(const row_label_type& row_id, const col_label_type& col_id, unsigned short int value);
     void setIsSymmetric(bool isSymmetric);
-    void setRows(unsigned int rows);
-    void setCols(unsigned int cols);
-    void setRowByRowLoc(const std::vector<unsigned short int> &row_data, unsigned int row_loc);
+    void setRows(unsigned long int rows);
+    void setCols(unsigned long int cols);
+    void setRowByRowLoc(const std::vector<unsigned short int> &row_data, unsigned long int row_loc);
     void setRowIds(const std::vector<row_label_type>& row_ids);
     void setColIds(const std::vector<col_label_type>& col_ids);
-    unsigned int addToRowIndex(const row_label_type& row_id);
-    unsigned int addToColIndex(const col_label_type& col_id);
-    void setDatasetRow(const std::vector<unsigned short int>& datasetRow, unsigned int row);
+    unsigned long int addToRowIndex(const row_label_type& row_id);
+    unsigned long int addToColIndex(const col_label_type& col_id);
+    void setDatasetRow(const std::vector<unsigned short int>& datasetRow, unsigned long int row);
     void setDataset(const std::vector<std::vector<unsigned short int>>& dataset);
 
     // Input/Output:
@@ -51,7 +51,7 @@ public:
     void printDataFrame() const;
 
     // Utility
-    bool isUnderDiagonal(unsigned int row_loc, unsigned int col_loc) const;
+    bool isUnderDiagonal(unsigned long int row_loc, unsigned long int col_loc) const;
 
 
 private:
@@ -60,15 +60,15 @@ private:
 
     // Private Members
     bool isSymmetric;
-    unsigned int rows;
-    unsigned int cols;
+    unsigned long int rows;
+    unsigned long int cols;
     std::vector<row_label_type> rowIds;
     std::vector<col_label_type> colIds;
 
-    std::unordered_map<row_label_type, unsigned int> rowIdsToLoc;
-    std::unordered_map<col_label_type, unsigned int> colIdsToLoc;
+    std::unordered_map<row_label_type, unsigned long int> rowIdsToLoc;
+    std::unordered_map<col_label_type, unsigned long int> colIdsToLoc;
 
     // Utility
-    unsigned int symmetricEquivalentLoc(unsigned int row_loc, unsigned int col_loc) const;
-    unsigned int dataset_size;
+    unsigned long int symmetricEquivalentLoc(unsigned long int row_loc, unsigned long int col_loc) const;
+    unsigned long int dataset_size;
 };
