+++ spatial_access/src/transitMatrix.h
@@ -16,7 +16,7 @@ void calculateRow(const std::vector<int> &dist, graphWorkerArgs<row_label_type, 
 /* Calulates distances of shortest paths from src to all*/
 /* vertices. It is a O(ELogV) function*/
 template<class row_label_type, class col_label_type>
-void dijkstra(unsigned int src, graphWorkerArgs<row_label_type, col_label_type> *wa);
+void dijkstra(unsigned long int src, graphWorkerArgs<row_label_type, col_label_type> *wa);
 
 template<class row_label_type, class col_label_type>
 void graphWorkerHandler(graphWorkerArgs<row_label_type, col_label_type>* wa);
@@ -35,13 +35,13 @@ public:
     int numNodes;
 
     // Constructors
-    transitMatrix(bool isSymmetric, unsigned int rows, unsigned int cols) : df(isSymmetric, rows, cols), numNodes(0) {}
+    transitMatrix(bool isSymmetric, unsigned long int rows, unsigned long int cols) : df(isSymmetric, rows, cols), numNodes(0) {}
 
     // Initialization
-    void prepareGraphWithVertices(unsigned int V);
-    void addToUserSourceDataContainer(unsigned int networkNodeId, const row_label_type& row_id, unsigned short int lastMileDistance);
-    void addToUserDestDataContainer(unsigned int networkNodeId, const col_label_type& col_id, unsigned short int lastMileDistance);
-    void addEdgeToGraph(unsigned int src, unsigned int dest, unsigned short int weight, bool isBidirectional);
+    void prepareGraphWithVertices(unsigned long int V);
+    void addToUserSourceDataContainer(unsigned long int networkNodeId, const row_label_type& row_id, unsigned short int lastMileDistance);
+    void addToUserDestDataContainer(unsigned long int networkNodeId, const col_label_type& col_id, unsigned short int lastMileDistance);
+    void addEdgeToGraph(unsigned long int src, unsigned long int dest, unsigned short int weight, bool isBidirectional);
     void addToCategoryMap(const col_label_type& dest_id, const std::string& category);
 
 
@@ -49,8 +49,8 @@ public:
     void compute(unsigned int numThreads);
     const std::vector<std::pair<col_label_type, unsigned short int>> getValuesBySource(row_label_type source_id, bool sort) const;
     const std::vector<std::pair<row_label_type, unsigned short int>> getValuesByDest(col_label_type dest_id, bool sort) const;
-    const std::unordered_map<row_label_type, std::vector<col_label_type>> getDestsInRange(unsigned int range, int numThreads) const;
-    const std::unordered_map<col_label_type, std::vector<row_label_type>> getSourcesInRange(unsigned int range, int numThreads) const;
+    const std::unordered_map<row_label_type, std::vector<col_label_type>> getDestsInRange(unsigned int range, unsigned int numThreads) const;
+    const std::unordered_map<col_label_type, std::vector<row_label_type>> getSourcesInRange(unsigned int range, unsigned int numThreads) const;
     unsigned short int timeToNearestDestPerCategory(const row_label_type& source_id, const std::string& category) const;
     unsigned short int countDestsInRangePerCategory(const row_label_type& source_id, const std::string& category, unsigned short int range) const;
     unsigned short int timeToNearestDest(const row_label_type& source_id) const;
@@ -58,17 +58,17 @@ public:
 
     // Getters
     unsigned short int getValueById(const row_label_type& source_id, const col_label_type& dest_id) const;
-    unsigned int getRows() const;
-    unsigned int getCols() const;
+    unsigned long int getRows() const;
+    unsigned long int getCols() const;
     bool getIsSymmetric() const;
-    const std::vector<unsigned short int>& getDatasetRow(unsigned int row) const;
+    const std::vector<unsigned short int>& getDatasetRow(unsigned long int row) const;
     const std::vector<std::vector<unsigned short int>>& getDataset() const;
     const std::vector<row_label_type>& getPrimaryDatasetIds() const;
     const std::vector<col_label_type>& getSecondaryDatasetIds() const;
 
     // Setters
-    void setRows(unsigned int rows);
-    void setCols(unsigned int columns);
+    void setRows(unsigned long int rows);
+    void setCols(unsigned long int columns);
     void setIsSymmetric(bool isSymmetric);
     void setDataset(const std::vector<std::vector<unsigned short int>>& dataset);
     void setPrimaryDatasetIds(const std::vector<row_label_type>& primaryDatasetIds);
